将远程附件抓取移到中立 Core 原子
